% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{al_sample}
\alias{al_sample}
\title{Length at age of an imaginary fish population}
\format{
A data frame with 400 rows and 2 variables:
\describe{
\item{Length}{Length, in centimeter}
\item{Age}{Age, in year}
}
}
\usage{
al_sample
}
\description{
A subsampled data of imaginary fish population with 400 individuals
}
\keyword{datasets}
