% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foward_alk.R
\name{apply_forward_alk}
\alias{apply_forward_alk}
\title{Apply forward alk to length frequency data}
\usage{
apply_forward_alk(aldata, lengthdata)
}
\arguments{
\item{aldata}{tibble with age and length data with
colnames = c("Length", "Age")}

\item{lengthdata}{tibble with length and frequency data with
colnames = c("Length", "Frequency")}
}
\value{
tibble with alk applied to the frequency data
}
\description{
Apply forward alk to length frequency data
}
\examples{
\dontrun{
mydata <- sampledata
mylfreq <- sampledata \%>\% dplyr::select(Length) \%>\% count_ldata(.)

apply_forward_alk(mydata, mylfreq)
}
}
