% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{categorize_num}
\alias{categorize_num}
\title{Categorize numerical vector to size or age class}
\usage{
categorize_num(data, interval = 1)
}
\arguments{
\item{data}{numeric vector}

\item{interval}{interval of size or age class}
}
\value{
numeric vector
}
\description{
Categorize numerical vector to size or age class
}
\examples{
\dontrun{
categorize_num(1:10, 1)
categorize_num(runif(10), 0.1)
}
}
