% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count.ldata}
\alias{count.ldata}
\title{Create length frequency data}
\usage{
count.ldata(dat)
}
\arguments{
\item{dat}{single strand vector with length data}
}
\value{
length frequency
}
\description{
Create length frequency data
}
\examples{
\dontrun{
ldata <- tibble::tibble(Length = round(runif(1000, 5, 40)))
count.ldata(ldata)
}
}
