% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foward_alk.R
\name{create_forward_alk}
\alias{create_forward_alk}
\title{Create forward alk}
\usage{
create_forward_alk(dat)
}
\arguments{
\item{dat}{tibble with two columns with colnames = c("Length", "Age")}
}
\value{
ALK data with n (frequency) and prop (proportion)
}
\description{
Create forward alk
}
\examples{
\dontrun{
create_forward_alk(sampledata)
}
}
