% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_alk.R
\name{create_inverse_alk}
\alias{create_inverse_alk}
\title{Create inverse alk}
\usage{
create_inverse_alk(dat)
}
\arguments{
\item{dat}{tibble with two columns with colnames = c("Length", "Age")}
}
\value{
ALK data with n (frequency) and prop (proportion)
}
\description{
Create inverse alk
}
\examples{
\dontrun{
create_inverse_alk(sampledata)
}
}
