% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plus_group.R
\name{create_plus_group}
\alias{create_plus_group}
\title{Create plus group of age}
\usage{
create_plus_group(age, plus_group)
}
\arguments{
\item{age}{Age data}

\item{plus_group}{integer age which floors upper age as plus group}
}
\description{
Create plus group of age
}
