% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk.R
\name{create_reverse_alk}
\alias{create_reverse_alk}
\title{Create reverse alk}
\usage{
create_reverse_alk(dat)
}
\arguments{
\item{dat}{tibble with two columns with colnames = c("Length", "Age")}
}
\value{
ALK data with n (frequency) and prop (proportion)
}
\description{
Create reverse alk
}
\examples{
\dontrun{
create_reverse_alk(sampledata)
}
}
