% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk.R
\name{forward_alk}
\alias{forward_alk}
\title{Create ALK by length}
\usage{
forward_alk(dat)
}
\arguments{
\item{dat}{tibble with two columns with colnames = c("Length", "Age")}
}
\value{
ALK data with n (frequency) and prop (proportion)
}
\description{
Create ALK by length
}
\examples{
\dontrun{
forward_alk(sampledata)
}
}
