% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freq_sample}
\alias{freq_sample}
\title{Length frequency data of an imaginary fish population}
\format{
A data frame with 2000 rows and 1 variables:
\describe{
\item{Length}{Length, in centimeter}
}
}
\usage{
freq_sample
}
\description{
A data set of imaginary fish population with 2000 individuals
}
\keyword{datasets}
