% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_length_bin.R
\name{limit_length_bin}
\alias{limit_length_bin}
\title{Limit length bin by applying plus and minus group}
\usage{
limit_length_bin(length_data, limit_length = c(NA, NA))
}
\arguments{
\item{length_data}{Length frequency data}

\item{limit_length}{Limit of minus and plus group
given as \code{c(lower_limit, upper_limit)}
\describe{
\item{lower_limit}{lower limit of length bin}
\item{upper_limit}{upper limit of length bin}
}}
}
\description{
Limit length bin by applying plus and minus group
}
\examples{
\dontrun{
limit_length_bin(Length = freq_sample, limit_length = c(10, 30))
}
}
