% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alk.R
\name{reverse_alk}
\alias{reverse_alk}
\title{Create ALK by age}
\usage{
reverse_alk(dat)
}
\arguments{
\item{dat}{tibble with two columns with colnames = c("Length", "Age")}
}
\value{
ALK data with n (frequency) and prop (proportion)
}
\description{
Create ALK by age
}
\examples{
\dontrun{
reverse_alk(sampledata)
}
}
