% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Length at age of an imaginary fish population}
\format{
A data frame with 2143 rows and 2 variables:
\describe{
\item{Length}{Length, in centimeter}
\item{Age}{Age, in year}
}
}
\usage{
sampledata
}
\description{
A dataset of imaginary fish population with over 2000 individuals
}
\keyword{datasets}
