% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsample_data}
\alias{subsample_data}
\title{Subsample sampledata}
\usage{
subsample_data(dat, sub_prop = 0.25)
}
\arguments{
\item{dat}{sampledata}

\item{sub_prop}{Proportion of data equipped with age}
}
\value{
sampledata with and without age
}
\description{
Subsample sampledata
}
\examples{
\dontrun{
subsample_data(subsample, sub_prop = 0.5)
}
}
